% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lmPredict}
\alias{lmPredict}
\title{lmPredict}
\usage{
lmPredict(x)
}
\arguments{
\item{x}{A data.frame of x values.}
}
\value{
y The prediction from the lm model.
}
\description{
A simple function loading a pre-packaged lm model.
}

