% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simpleRandomPrediction}
\alias{simpleRandomPrediction}
\title{simpleRandomPrediction}
\usage{
simpleRandomPrediction(x)
}
\arguments{
\item{x}{The upper boundary}
}
\value{
A number between 1 and x
}
\description{
A simple function that samples a number between 1 and x.
}

